Implement experimental high-memory bootloader
The bootloader now lives at the high end of flash, allowing an non-
bootloader-aware application to run without modifications.

The bootloader modifies the application's vector table to point to
the bootloader's reset stub. The actual reset vector is stored in
unused entries in the application's vector table.

Based on ideas from:
http://stm32duino.com/viewtopic.php?t=687

Since the bootloader's vector table is now shared with the application,
there is now also enough space to fit the DFU upload command in 8KiB.