dfu: support "detach"
When in DFU mode, the only way to exit was via a download operation.
"dfu-util -e" to request a detach was simply not supported, so the
bootloader simply remains sitting there
Handle detach via reboot, the usb layer will re-enumerate if it can,
and will boot app/bl based on it's own decisions.

Signed-off-by: Karl Palsson <karlp@tweak.net.au>