DEMO: Support STM32L0
Demonstration of an m0+ target, and the ease of porting.

Unfortunately, some internal library code uses a modulo operation, and
that pulls in software divide, and this blows out past the 8k assumed
bootloader size.  You can simply comment out

  webusb_setup(usbd_dev);
  winusb_setup(usbd_dev);

in dapboot.c and it will drop under 8k again, but's clearly not a viable
long term solution.