Fix detach request callback being treated as manifest request callback
This prevents the bootloader from detaching prematurely after download and
makes it compliant with the DFU 1.1 spec. Many flashing routines do not
expect an automatic detach, so I've assumed the DFU spec option to return
to idle after download as what was originally intended (an automatic reset
can still be achieved by passing the -R flag to dfu-util for example), in
other words the bootloader was intended to have the USB_DFU_MANIFEST_TOLERANT
attribute since there is no STATE_DFU_MANIFEST_WAIT_RESET in the code.