Fix RTC backup register indexing, add target-specific register and boot command
According to ST's RM0008 STM32F103xx reference manual section 6.4.5 table 17
the backup registers are 32 bits wide with only the lower 16 bits being used
for storing values. In addition to this, the "first" register slot is entirely
reserved and there is no backup register zero, meaning that BKP_DR1 starts from
an offset of 0x04. This was mostly correctly reflected in backup.c, but the
2-multiplier for the register enum throws everything off by a lot, which
results in the boot command check failing to work correctly when using anything
but the first backup register. BKP0 has also been removed since as described by
RM0008 indexing starts at one (BKP1 is now fully equivalent to the old BKP0).

I've also gone ahead and implemented support for specifying the backup register
and boot command to use per-target. If not defined, options equivalent to the
old behavior are used to not break existing applications.