Enable specifying the alternate setting for the DFU interface per-target
This works by "padding" the DFU alternate setting with dummy no-op descriptors
that occupy all previous alternate setting values. AFAICT while these no-op
interfaces are not documented in the USB defined class codes, they don't
identify as anything in Linux and cause no driver to be loaded nor any warnings
in the system logs, and they should be fine for other operating systems as
well. When USB_ALT > 0, the DFU interface appears in the specific alternate
setting, which dfu-util does automatically identify and use, alternatively it
can be manually targeted with the -a flag. To have USB_ALT be a per-target
(optional) property the target.h implementor must load it (since it has the
macro value to keep everything compile-time), which is why I've had to split
out the descriptor structs into their own file to be preprocessed as part of
target_stm32f103.c or target_stm32l1.c. The USB initialization code can then
call target_usb_descriptor() to get the pointer to the configuration
descriptor defined by the target.