Add support for the BigTreeTech SKR MINI E3 V2.0 board and STM32F103RC
The BigTreeTech SKR MINI E3 V2.0 is a 3D printer motherboard based on the
STM32F103RC with 256 KiB of flash. Out of the box it has a MOSFET controlling
the USB pullup wired to the SWCLK pin, which makes attaching a debugger very
difficult. The USBMOD target variant has pullup control disabled for people
that feel adventurous enough to desolder some components from their board to
fix this issue, there is a more in-depth explanation in the USBMOD config.h.
Both targets have a high memory variant which has been tested and confirmed
working with the Klipper firmware by specifying "no bootloader" in its memory
offset configuration menu.