Add DFU_WILL_DETACH option for automatic detach/reset and enable it by default
Enabling DFU_WILL_DETACH will set bitManifestationTolerant = 0 and
bitWillDetach = 1, disabling it will set the opposite. If DFU_WILL_DETACH
is set, the target will automatically detach and reset following a successful
reprogramming, and if it's not set, the target will enter the dfuIDLE state.
Both of these behaviors are described in the DFU 1.1 spec. DFU_WILL_DETACH is
enabled by default for WebUSB compatibility, as some platforms will apparently
not instruct the target to reset after flashing.