Detect USB requests and reset or enter error state
Per the DFU spec, a USB reset should cause the device to either reenumerate as the application or enter an error state if invalid.
Register a USB reset callback to detect and handle USB resets, taking care not to trigger the first time the bootloader is enumerated.
Run the manifest callback if present when manifesting via USB reset.